Update libvpx to version 1.6.0
This invalidates a set of configurations because legal values for
VP9 "cpu-used" have changed, and require an update to the fixed-q
patch for VP8.

This touches a number of verification tools.

NOTE this commit: Verification tool work showed that the way
"verify_scores" works is not good. Another CL intended in order
to fix that problem.